Change type of rating in blog
bug on postgre

+++ db/migrate/20131218221704_add_rating_to_blogs.rb
@@ -1,5 +1,5 @@
 class AddRatingToBlogs < ActiveRecord::Migration
   def change
-    add_column :blogs, :rating, :double
+    add_column :blogs, :rating, :float
   end
 end
